LFList split into cpp
